Online contact book: initial commit
This is a simple online contact book that implemented in Java platform. Main aim of this project is to use hibernate, spring and lombok frameworks. The project used servlets container and get better understanding of how it works. The online contact book is based on TDD principles, what is important by "clean code" development.